code(fix): set frame_skip as an int

+++ B-PROS(No-ops)/mainBpro.cpp
@@ -56,7 +56,7 @@ int main(int argc, char** argv){
 
 	ale.setFloat("repeat_action_probability", 0.00);
 	ale.setInt("random_seed", 2*param.getSeed());
-	ale.setFloat("frame_skip", param.getNumStepsPerAction());
+	ale.setInt("frame_skip", param.getNumStepsPerAction());
 	ale.setInt("max_num_frames_per_episode", param.getEpisodeLength());
     ale.setBool("color_averaging", true);
